Construct covariance matrix (K) for given data and hyperparameters

This function is mostly used internally. In most common use, user will call gpFit() or predict(), which call this function themselves.

K_matrix(
  gp,
  hyperpar = gpHyperparList(gp),
  x1 = gp$data,
  x2 = NULL,
  K.cache = NULL,
  components = NULL,
  comp_means = FALSE,
  comp_means_weights = NULL
)

Arguments

gp

object of class gp.

hyperpar

(optional) lists of hyperparameter values, as returned by gpHyperparList(); default is to use the hyperparameters from the gp object.

x1

object of class gpData - data corresponding to rows of the resultant covariance matrix. It is supposed to be correctly scaled already! (see gpDataPrepare())

x2

(optional) object of class gpData - data corresponding to columns of the resultant covariance matrix. If not supplied, x1 is used and the covariance matrix is square and symmetric. It is supposed to be correctly scaled already! (see gpDataPrepare())

K.cache

(optional) object returned by K_cache() function, to speed up repeated calls to K_matrix() and dK_dhi() functions.

components

of the covariance formula. One of the following:

NULL : all components will be selected (the default)

character vector: components to be selected

formula: additive formula specifying components, same syntax as in update.formula, e.g.

~. 	all components

~.-env-spat 	all components except env and spat

~spat+year 	only components spat and year

comp_means

only meaningful for x1 being the training dataset and x2 = NULL. If set to TRUE, then for each component, calculate vector of covariances of the training data with the average of all the predictions from the training dataset, for that particular component. These vectors are the rowmeans of training covariance matrices (that's why called component means). Save them in an attribute comp_means of K. These are later used in the predict method.

For comp_means = TRUE, x1 should be the training dataset. Perhaps it might make sense in other cases as well, but just make sure you know what you're doing :)

comp_means_weights

don't have to sum up to 1, will be normalized

Value

Covariance matrix K:

Number of rows will be equal to gpDataSize(x1, gp$GP_factor), where gp$GP_factor is the grouping factor corresponding to the dimension of the Gaussian process.

Analogically, if x2 is supplied, the number of columns will be equal to gpDataSize(x2, gp$GP_factor).

If x2 is not supplied, the covariance matrix is square and symmetric, representing the covariances among all rows in x1.