Construct gaussian process model

Construct gaussian process model

gp(
  f,
  data,
  lik,
  lik.hyperpar = NULL,
  lik.formula = NULL,
  lik.reindex2main = TRUE,
  W.type = c("diag", "bdiag", "other"),
  predictor.fun = NULL,
  response.parname = NULL,
  link = NULL
)

Arguments

f

formula for the covariance matrix

data

object of class gpData, created by the gpData() function

lik

an object of class likTemp, representing the user-defined template of the likelihood function p(y|f).

ORIG DOC TO BE DELETED - MAYBE USE PART OF THESE FOR THE DOCUMENTATION OF CLASS likTemp either a single function, or a named list of three functions (three-stage template); these are two ways to supply the user-defined negative log likelihood function p(y|f):

negLogLik is a single function f(data, par), taking two parameters:

data

object of class gpData, containing both predictors and a response variable

par

a list of parameters (numeric scalars or vectors). Always includes f, a numeric vector - result of the gaussian process. The par$f vector can be either of the dimension of the gaussian process, or the main table - see the negLogLik.reindex2main parameter. The par parameters list can optionally include other parameters, see the negLogLik.hyperpar argument below.

Logic is similar as in the previous case, but the negLogLik is supplied as a chain of three consecutive functions (three-stage template). Then, negLogLik is a named list of three functions:

pred.link

either character or function(data, par). This is predictor & link component. Processes the parameters into other parameters; is also used by predict(type = "response") for predictions, therefore, the data might not contain the response variable (thus, the response variable must not be used by this function). Should be either a character string representing the name of a link function (this will be the most frequent case; see !!! for supported link functions), or a function(data, par) (see above for parameter description !!!). This function will return a named list of parameters par that will be passed on to the next stage.

process

function(data, par), this function processes the parameters and the data (which always contain the response variable here) and shapes both parameters and data for particular likelihood function. This function should return a named list required by particular likelihood function (see !!!). This component is used by cross-validation and Laplace-Fisher approximation.

lik

function evaluating the likelihood. Either character representing name of likelihood function (see !!! for supported values), or custom function(x) taking x from the previous stage.

lik.hyperpar

optional, a named list of extra likelihood hyperparameters for the negLogLik function (named list of numeric scalars or vectors); this doesn't include the special parameter f (the Gaussian process). This effectively defines the likelihood hyperparameters, named ".lik" in the hyperparameter table.

lik.formula

optional, formula related to the supplied negLogLik function. This is not used in any way, except it is stored in the object and printed in the summary, so it has only informational purpose.

lik.reindex2main

should the par$f parameter of the negLogLik function be reindexed to the rows of the main table? If FALSE, the par$f parameter will be kept at the dimension of the Gaussian Process (as reported by gpDetermineSize()). If the Gaussian process is running at the dimension of the main table, then it does not matter.

W.type

the type of the hessian matrix (W) of the negative log likelihood function (as defined by the lik parameter)

"diag"

(default) - the hessian matrix (W) is diagonal. Most common case, fastest computation. This is usually the case when the user-defined likelihood factorizes over individual \(f_i\)'s.

"bdiag"

the hessian matrix (W) is block-diagonal. This is usually the case when the user-defined likelihood factorizes over mutually disjoint sets of \(f_i\)'s.

"other"

the hessian matrix (W) is a general matrix.

One should keep the setting as specific as possible to keep the computation efficient.

Value

object of class gp - the gaussian process model. Some important slots:

$hyperpar

hyperparameter table (see XXX). User is invited to tweak these settings.

$obsdata

observed training data, in the original scale, as user supplied them

$data

training data prepared for model fit - converted to matrices, and scaled (standardized) where appropriate