Fit a Gaussian process model

Fit a Gaussian process model

gpFit(
  gp,
  h = NULL,
  opt.h = TRUE,
  two.stage = TRUE,
  stages = 1:2,
  stage1low = 0.01,
  use.prior = TRUE,
  hessian = FALSE,
  opt.control = list(factr = 1e+10, maxit = 190),
  verbose = FALSE,
  method = c("Laplace"),
  use_f_start = TRUE,
  weights = NULL,
  grad.computation = TRUE,
  warn.exclude = NULL,
  recursive = FALSE
)

Arguments

gp

GP model object.

h

optional; a numeric vector of starting values of hyperparameters (untransformed, i.e. on their most natural scale). If NULL, the starting values are taken from gp hyperparameter table (gp$hyperpar[,"start"]).

opt.h

logical; should hyperparameters be optimised? Default TRUE; Only rarely you may want to say FALSE.

two.stage

logical; if TRUE (the default), the two-stage optimization method will be applied (see Details below).

stages

numerical vector - the optimization stages to be performed. Only relevant for two.stage = TRUE. Default is 1:2, i.e. both stages. See details below.

If you are running only stage 2 separately, which would make sense only in rare cases (perhaps two.stage = FALSE is what you need in most of those cases), note that stage 2 doesn't start from gp$hyperpar$start, but from gp$hyperpar$value.

stage1low

the low threshold for the sigma2 parameter of the components for which the two-stage fitting applies.

use.prior

logical; if TRUE, likelihood + prior (posterior) is optimized; if FALSE, only likelihood is optimized.

hessian

logical; compute and return the Hessian during the optimization. Warning: may be heavy on CPU time! Default FALSE.

opt.control

list of control parameters passed to the optim(control = )

verbose

logical

method

character; fitting method. Currently "Laplace" supported (Laplace approximation).

use_f_start

logical; internal optimization - start each Laplace approximation from the optimized f vector from the previous hyperparameter iteration.

weights

numeric or NULL; observation weights (not implemented).

grad.computation

logical; compute gradients during the gaussian process optimization (required when opt.h = TRUE).

warn.exclude

character vector; vector of warnings that should be disabled. Allowed values are: "warn.stage2.start". Default NULL.

recursive

internal option; do not use.

Value

Returns the input gp object with the model fit (stored in gp$fit) and updated hyperparameters.

NOTE!! The returned object is usually large, because it contains pre-calculated matrices that are used for predictions. If you need to save the model, call gpPack() first to remove these; when you need to use the model again, call gpUnpack() to restore these matrices.

Details

The two-stage fitting method (two.stage = TRUE) is a special optimization procedure (Telenský et al 2026) where the optimization is split into two stages:

In the first stage, sigma2 parameters are not allowed to go too close to zero (as per the stage1low parameter). This way, the other hyper-parameters in the same component have time to optimize themselves before the sigma2 parameter could shut the whole component down to zero. (This is only applied to components that have also other hyper-parameters apart from sigma2.)

In the second stage, the restriction applied in the first stage is lifted and all the sigma2 hyper-parameters are allowed to go further down to zero.

In some cases, where user tries to optimize the CPU time by manually splitting the procedure in multiple steps, it might be a good idea to run both stages only in the last step, and use only stage 1 in all the previous steps (see Vignette XXX).

If you are running only stage 2 separately, which would make sense only in rare cases (perhaps two.stage = FALSE is what you need in most of those cases), note that stage 2 doesn't start from gp$hyperpar$start, but from gp$hyperpar$value.

Examples

if (FALSE) { # \dontrun{
gp <- gp(...) # construct gp object
gp <- gpFit(gp)
} # }