Perform an N-fold cross-validation of a Gaussian process model

Function will fit the GP model for each of the N cross-validation folds, and evaluate cross-validation prediction for the whole training data set.

gpFitCV(
  gp,
  fold.col,
  fold.fact = "1",
  folds = NULL,
  start.from.model = NULL,
  parallel = TRUE,
  fn.prefix = "",
  log.fn = if (parallel) "log-fold%f-%h-%p.txt" else NULL,
  dump.fn = if (parallel) "dump-fold%f-%h_%p" else NULL,
  tr.max.lines = 5,
  ...
)

Arguments

gp

GP model object

fold.col

either a name of a column in the main table, or a vector along factor fold.fact. The column (or the supplied vector) must be a vector of integers from 1 to N (N being the number of folds), specifying the number of the cross-validation fold the given record belongs to.

fold.fact

a factor along which the cross-validation folds (fold.col) are specified. Factor "1" means the folds are specified for the rows of the main table. Note that if the GP dimension is given by some real grouping factor (i.e. p$GP_factor != "1"), then the fold.fact must be that factor.

folds

integer vector of folds to fit; if NULL (the default), all folds are fit

start.from.model

an object of class gp to take the starting values from. If specified, then for each cross-validation fold model to be fit, the starting values of the hyperparameters will be taken from the corresponding fold model wherever possible.

parallel

should the cross-validation models be run in parallel? Uses foreach(), and requires parallel background for foreach() to be already registered.

fn.prefix

character; prefix of file names for log files (log.fn) and dump files (dump.fn), in case these are not NULL

log.fn

if not NULL, standard and error output of each fold model job will be saved into a log file with this file name. Special % sequences can be used, see Details below.

dump.fn

if not NULL, debug dump of given fold model will be saved upon an error, with this file name (the .rda extension will be added to it). Special % sequences can be used, see Details below.

tr.max.lines

the max.lines parameter for traceback, i.e. the maximum number of lines printed per call when error occurs

...

options to be passed to the gpFit() method.

Details

The arguments log.fn and dump.fn allow for special sequences:

%f - number of the cross-validation fold

%h - hostname, i.e. the name of the machine where the worker job runs

%p - process ID of the worker job