Calculate the approximation of the LGO-CV (leave-group-out cross-validation) and evaluate it.

Works at the same level as gpFitCV - makes CV predictions and evaluates them to the gp object. There is an important difference though: LGO-CV works for a given set of hyperparameters, whereas the default gpFitCV() optimises them, and thus cross-validates also the hyperparameter optimization process.

gpLGO(
  gp,
  fold.col,
  fold.fact = "1",
  pred.null = NULL,
  lmFit.options = list(lik.hyperpar = "fix"),
  lmFit.pred.options = list(type = "terms"),
  hyperpar = gpHyperparList(gp),
  chunked = FALSE
)

Arguments

gp

GP model object

fold.col

either a name of a column in the main table, or a vector along factor fold.fact. The column (or the supplied vector) must be a vector of integers from 1 to N (N being the number of folds), specifying the number of the cross-validation fold the given record belongs to.

fold.fact

a factor along which the cross-validation folds (fold.col) are specified. Factor "1" means the folds are specified for the rows of the main table. Note that if the GP dimension is given by some real grouping factor (i.e. gp$GP_factor != "1"), then the fold.fact must be that factor.

pred.null

optional predictions from the null model (predict.lmFit(type = "terms")). If these are available from the actual exact lmFitCV call, it is better to provide them this way. Otherwise the null model predictions will be calculated just by fitting all data as training, which will result in a slightly better null model (and a slightly lower pseudo-R2 statistics then).

lmFit.options

list of options to be passed to lmFit when fitting the intercept-only null model. Only used when pred.null = NULL.

lmFit.pred.options

list of options to be passed to the predict.lmFit method when predicting from the null model. Only used when pred.null = NULL.

hyperpar

(optional) lists of hyperparameter values, as returned by gpHyperparList(); default is to use the hyperparameters from the gp object.

chunked

logical, passed to gpLGOpred; use dense fold blocks? This currently requires diagonal W.

Value

the gp model object with approximate LGO-CV predictions and statistics in gp$fit$LGOCV.