Inspect warnings and their call stacks

Runs an expression and prints diagnostic information for each warning: the warning message, condition call, and R-visible call stack. Warnings are not suppressed.

inspect_warnings(expr, stack = c("condition", "full"))

Arguments

expr

Expression to evaluate.

stack

Character scalar. "condition" prints the stack only up to the warning's conditionCall(). "full" prints the full R-visible stack, including handler internals.

Value

The value of expr, invisibly if expr returns invisibly. Called for its warning diagnostics side effect.

Examples

f <- function() {
  warning("f warned")
}

inspect_warnings(f(), stack = "condition")
#> 
#> --- warning inspection ---
#> message: f warned
#> condition call: f()
#>  frame     package
#>      1        base
#>      2        <NA>
#>      3        <NA>
#>      4        <NA>
#>      5        <NA>
#>      6        <NA>
#>      7        <NA>
#>      8        base
#>      9        base
#>     10        base
#>     11        base
#>     12 R_GlobalEnv
#>     13     pkgdown
#>     14     pkgdown
#>     15     pkgdown
#>     16     pkgdown
#>     17        base
#>     18       purrr
#>     19       purrr
#>     20       purrr
#>     21        base
#>     22       purrr
#>     23     pkgdown
#>     24        base
#>     25     pkgdown
#>     26     pkgdown
#>     27     pkgdown
#>     28     downlit
#>     29    evaluate
#>     30        base
#>     31        <NA>
#>     32        <NA>
#>     33        <NA>
#>     34        <NA>
#>     35        <NA>
#>     36        <NA>
#>     37    evaluate
#>     38        base
#>     39        <NA>
#>     40        base
#>     41        base
#>     42        base
#>     43        <NA>
#>     44          gp
#>     45        base
#>     46        <NA>
#>                                                            fun
#>                                                 base::tryCatch
#>                                                   tryCatchList
#>                                                    tryCatchOne
#>                                                     doTryCatch
#>                                                   tryCatchList
#>                                                    tryCatchOne
#>                                                     doTryCatch
#>                                      base::withCallingHandlers
#>                                                  base::saveRDS
#>                                                  base::do.call
#>  function (what, args, quote = FALSE, envir = parent.frame()) 
#>                        function (..., cli_colors, hyperlinks) 
#>                                            pkgdown::build_site
#>                                               build_site_local
#>                                                build_reference
#>                                             unwrap_purrr_error
#>                                            withCallingHandlers
#>                                                     purrr::map
#>                                                           map_
#>                                            with_indexed_errors
#>                                            withCallingHandlers
#>                                              call_with_cleanup
#>                                                             .f
#>                                            withCallingHandlers
#>                                           data_reference_topic
#>                                                   run_examples
#>                                             highlight_examples
#>                                downlit::evaluate_and_highlight
#>                                             evaluate::evaluate
#>                                                   withRestarts
#>                                                withRestartList
#>                                                 withOneRestart
#>                                               doWithOneRestart
#>                                                withRestartList
#>                                                 withOneRestart
#>                                               doWithOneRestart
#>                                                  with_handlers
#>                                                           eval
#>                                                           eval
#>                                            withCallingHandlers
#>                                                    withVisible
#>                                                           eval
#>                                                           eval
#>                                               inspect_warnings
#>                                            withCallingHandlers
#>                                                              f
#>                                                                                                                                                            call
#>                                                                                                                      base::tryCatch(base::withCallingHandlers({
#>                                                                                                                tryCatchList(expr, classes, parentenv, handlers)
#>                                                                                          tryCatchOne(tryCatchList(expr, names[-nh], parentenv, handlers[-nh]), 
#>                                                                                                              doTryCatch(return(expr), name, parentenv, handler)
#>                                                                                                        tryCatchList(expr, names[-nh], parentenv, handlers[-nh])
#>                                                                                                             tryCatchOne(expr, names, parentenv, handlers[[1L]])
#>                                                                                                              doTryCatch(return(expr), name, parentenv, handler)
#>                                                                                                                                     base::withCallingHandlers({
#>  base::saveRDS(base::do.call(base::do.call, base::c(base::readRDS("C:\\\\Users\\\\Tomas\\\\AppData\\\\Local\\\\Temp\\\\RtmpOGBu3x\\\\callr-fun-7c84736c1897"), 
#>                base::do.call(base::do.call, base::c(base::readRDS("C:\\\\Users\\\\Tomas\\\\AppData\\\\Local\\\\Temp\\\\RtmpOGBu3x\\\\callr-fun-7c84736c1897"), 
#>                                                                                                  (function (what, args, quote = FALSE, envir = parent.frame()) 
#>                                                                                                                        (function (..., cli_colors, hyperlinks) 
#>                                                                                                                                        pkgdown::build_site(...)
#>                                                                                  build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run, 
#>                                                                            build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run, 
#>                                                                                                   unwrap_purrr_error(purrr::map(topics, build_reference_topic, 
#>                                                                                                 withCallingHandlers(code, purrr_error_indexed = function(err) {
#>                                                                                              purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy, 
#>                                                                                                                map_("list", .x, .f, ..., .progress = .progress)
#>                                                                                      with_indexed_errors(i = i, names = names, error_call = .purrr_error_call, 
#>                                                                                                               withCallingHandlers(expr, error = function(cnd) {
#>                                                                                                   call_with_cleanup(map_impl, environment(), .type, .progress, 
#>                                                                                                                                                .f(.x[[i]], ...)
#>                                                                              withCallingHandlers(data_reference_topic(topic, pkg, examples_env = examples_env, 
#>                                                                                                  data_reference_topic(topic, pkg, examples_env = examples_env, 
#>                                                run_examples(tags$tag_examples[[1]], env = if (is.null(examples_env)) NULL else new.env(parent = examples_env), 
#>                                                                                                                      highlight_examples(code, topic, env = env)
#>                                                                                               downlit::evaluate_and_highlight(code, fig_save = fig_save_topic, 
#>                                                                    evaluate::evaluate(code, child_env(env), new_device = TRUE, output_handler = output_handler)
#>                                                                                                                                    withRestarts(with_handlers({
#>                                                                                                                                 withRestartList(expr, restarts)
#>                                                                                            withOneRestart(withRestartList(expr, restarts[-nr]), restarts[[nr]])
#>                                                                                                                         doWithOneRestart(return(expr), restart)
#>                                                                                                                            withRestartList(expr, restarts[-nr])
#>                                                                                                                            withOneRestart(expr, restarts[[1L]])
#>                                                                                                                         doWithOneRestart(return(expr), restart)
#>                                                                                                                                                 with_handlers({
#>                                                                                                                                                      eval(call)
#>                                                                                                                                                      eval(call)
#>                                                                                                             withCallingHandlers(code, message = function (cnd) 
#>                                                                                                                                  withVisible(eval(expr, envir))
#>                                                                                                                                               eval(expr, envir)
#>                                                                                                                                               eval(expr, envir)
#>                                                                                                                      inspect_warnings(f(), stack = "condition")
#>                                                                                                               withCallingHandlers(expr, warning = function(w) {
#>                                                                                                                                                             f()
#> --- end inspection ---
#> 
#> Warning: f warned