Fit a linear model on the GP model object

Fits a linear model, where the latent Gaussian process f is replaced by a linear predictor constructed from formula.

lmFit(
  gp,
  formula,
  data,
  table = NULL,
  beta.name = "beta_lm",
  lik.hyperpar = c("fix", "optimize"),
  lik.fix = NULL,
  lik.prefix = ".lik.",
  silent = TRUE,
  optCtrl = list(iter.max = 300, eval.max = 400)
)

Arguments

gp

gp model object; its likelihood template will be used

formula

the linear model formula to replace the latent Gaussian process f

data

object of class gpData; data used for the likelihood and the linear predictor. See the details below.

table

character or NULL - name of the table in data that is going to be used to construct the model matrix for the linear model. This table must be of the dimension of gp$GP_factor. For the null model (~ 1), table can be NULL.

beta.name

character; name of the linear predictor coefficient parameter

lik.hyperpar

character; if there are likelihood hyperparameters, should they be fixed or optimized?

"fix"

the likelihood hyperparameters will be fixed at value specified by lik.fix

"optimize"

the likelihood hyperparameters will be optimized along with the parameters in formula

lik.fix

when lik.hyperpar = "fix", to what value should the likelihood hyperparameters be fixed?

single numeric value

all likelihood hyperparameters will be set to this value

"value"

the likelihood hyperparameters will be set to values in the column "value" in the hyperparameter table (gp$hyperpar$value)

lik.prefix

character; prefix added to likelihood hyperparameter names in summary()

silent

logical; passed to RTMB::MakeADFun

optCtrl

list; passed to nlminb(control = )

Value

Object of class lmFit

Details

Note that the model is linear in the formula specified. If lik.hyperpar = "optimize", the model is not necessarily linear in these hyperparameters; this is given by the likelihood template defined in the gp object.

The table used for the linear model must have the same dimension as gp$GP_factor, because the linear predictor replaces the latent Gaussian process f directly. If gp$GP_factor == "1", the table must have no grouping factor. For an intercept-only formula, lmFit() constructs the model matrix at the gp$GP_factor dimension directly, so no table needs to be supplied.

Note that no scaling (covariate standardization) is done neither in this function nor in predict. It is up to the user to provide the correct data argument with the desired scaling.

This function is used in the cross-validation process to fit the null model for the baseline comparison.

RTMB engine is used to fit the linear model.