Predict from a fitted Gaussian process model

Predict from a fitted Gaussian process model.

# S3 method for class 'gp'
predict(
  gp,
  newdata = NULL,
  type = c("latent", "terms", "response"),
  se.fit = FALSE,
  cov.fit = FALSE,
  conf.int = FALSE,
  conf.level = 0.95,
  components = NULL,
  comp_missing = c("avg", "none"),
  w = NULL,
  groupMeans = NULL,
  hyperpar = gpHyperparList(gp),
  maxn = Inf,
  parallel = FALSE,
  log.fn = NULL,
  dump.fn = NULL,
  log.append = FALSE,
  tr.max.lines = 5,
  pred.sims = 1e+05,
  Kx.cache = NULL,
  Kxx.cache = NULL,
  ...
)

Arguments

gp

object of class gp, the fitted Gaussian process model

newdata

object of class gpData, unscaled (i.e. not passed through gpDataPrepare()) data! They must be on the same scale as the unscaled training dataset used for fitting the model. If NULL (the default), predictions will be made on the training dataset (which is going to be super fast if conf.int = FALSE)

type

character; type of prediction:

"latent"

predictions directly from the Gaussian Process itself (f and optionally f_SE), always in the dimension of the GP (gp$GP_factor)

"terms"

predictions passed through first stage of the likelihood template, that is, including GP (f and optionally f_SE) along with all the other terms used in the likelihood template. If gp$lik.reindex2main is TRUE, the predictions will be reindexed to the dimension of the main table.

"response"

predictions passed through the first two stages of the likelihood template, that is, after applying the inverse link function. If gp$lik.reindex2main is TRUE, the predictions will be reindexed to the dimension of the main table.

cov.fit

logical; if TRUE, the covariance matrix of the latent f (the Gaussian process) will also be returned, in the original dimension of the GP. Note that the covariance matrix can be very memory consuming for large datasets. Default is FALSE.

conf.int

logical; should confidence intervals be calculated? If FALSE, no confidence intervals will be calculated, which makes the prediction much faster (but just in case both se.fit and cov.fit are also FALSE). If TRUE, se.fit will be automatically set to TRUE, since the standard errors are needed for the CI calculation.

conf.level

numeric; confidence level for the confidence intervals. Default is 0.95.

components

of the covariance formula. One of the following:

NULL : all components will be selected (the default)

character vector: components to be selected

formula: additive formula specifying components, same syntax as in update.formula, e.g.

~. 	all components

~.-env-spat 	all components except env and spat

~spat+year 	only components spat and year

comp_missing

how to compensate for missing components (if there are any), in mean and variance of the predictions:

"avg"

each omitted component is replaced by its posterior prediction for the average of that component over the training dataset. !!Overit: for these components, the variance/CI corresponds to confidence interval (uncertainty around the mean predictions) and not the prediction interval - see http://www.sthda.com/english/articles/40-regression-analysis/166-predict-in-r-model-predictions-and-confidence-intervals/ For models which have grouping factors, be aware of how the average is weighted - see CAUTION CAU001 in the code.

"none"

no compensation. In this case, the interpretation of the result's variance isn't much clear (!!!??? why?).

Thus, the most reasonable setting now is "avg"; unless one cares only about the prediction mean (f) of the latent variable, in which case a simple "none" would also do the job.

w

providing this optional argument will calculate average prediction(s) of all data points in this dataset, given these weights. Can be either a vector of the same length as the size of newdata (in which case a single average is calculated), or, if more different averages with different weights are desired, w can be a matrix where each column corresponds to one average - one set of weights (i.e. number of rows of w must be equal to the size of newdata, number of columns equal to the number of different averages to be calculated).

groupMeans

optional, factor of the size of the predicted dataset. If provided, mean predictions for each factor level are calculated.

hyperpar

(optional) lists of hyperparameter values, as returned by gpHyperparList(); default is to use the hyperparameters from the gp object.

maxn

maximum dataset size (along the gp$GP_factor) to fit at once; if the dataset is larger, it will be split into chunks of size maxn before predictions are calculated. Use Inf to disable the splitting entirely. Choosing a suitable value may speed up the computation when se.fit = TRUE and cov.fit = FALSE, because the Kxx matrices are then smaller (that's probably the only part where splitting helps) - but testing is needed.

The splitting is only available when newdata is not NULL, cov.fit = FALSE, and w, groupMeans, Kx.cache, and Kxx.cache are all NULL. (Note: splitting disabled for now by default, since in some models we would need to specify along which factor to split - cannot always use gp$GP_factor)

parallel

should prediction chunks (see maxn) be run in parallel? Requires default cluster to be already registered.

log.fn

if not NULL, standard and error output of each parallel prediction chunk job will be saved into a log file with this file name template. Special % sequences can be used, see Details below.

dump.fn

if not NULL, debug dump of a failed parallel prediction chunk job will be saved using this file name template. Special % sequences can be used, see Details below; the .rda extension is added after expansion.

log.append

logical; if TRUE, append to log.fn instead of overwriting it. Might be useful to create per-worker rather than per-job logs.

tr.max.lines

the max.lines parameter for traceback, i.e. the maximum number of lines printed per call when error occurs

pred.sims

(currently unused) number of simulations to be used for calculating the response scale predictions and their CIs. Default is 100000. Higher values give more accurate results, but are slower. Ignored if conf.int = FALSE.

Kx.cache

optional, object returned by K_cache() function, to speed up repeated calls to predict.gp() with the same newdata by caching parts of the K(training_data, newdata) matrix.

Kxx.cache

optional, object returned by K_cache() function, to speed up repeated calls to predict.gp() with the same newdata by caching parts of the K(newdata, newdata) matrix.

Value

A matrix of predictions, with columns f and f_SE (if se.fit = TRUE). If cov.fit = TRUE, returns a named list, pred will be the mentioned matrix and cov will be the full covariance matrix. For the dimension of the prediction, see the type parameter.

Details

The arguments log.fn and dump.fn are file name templates used when parallel = TRUE and predictions are split into chunks (see maxn. They allow for special sequences, expanded in the worker process:

%h - hostname, i.e. the name of the machine where the worker job runs

%p - process ID of the worker job You can use these in log.fn in conjunction with log.append = TRUE to create per-worker log files.