Redirect stdout and stderr to a log file

Redirects R's standard output and messages (stderr) to the specified file. On Unix systems, can optionally keep output visible on the console using tee and add timestamps via the ts command.

startLog(
  filename,
  tee = FALSE,
  all_or_nothing = TRUE,
  add_timestamp = FALSE,
  append = FALSE,
  autoWarn1 = TRUE
)

Arguments

filename

character. Path to the log file.

tee

logical. If TRUE, stream output to console as well (Unix only). currently on windows, STDERR cannot be split (i.e. will not be displayed on the console).

all_or_nothing

logical. Makes only sense for tee = TRUE. When TRUE, perform "tee" only on systems where full funcionality is available

add_timestamp

logical. Prepend timestamps using ts (Unix only, requires ts from moreutils).

append

logical. If TRUE, append to filename instead of overwriting.

autoWarn1

logical. If TRUE, it will set options(warn = 1) so that warnings are printed to the log immediately as well.

Value

A connection object used for logging, or NULL on failure.

Examples

if (FALSE) { # \dontrun{
LOG <- startLog("mylog.txt", tee = interactive(), append = TRUE)
# ... do work ...
endLog(LOG)
} # }