Make a gp object

Make a gp object

Usage

gp(rows = NULL, cols = NULL, wells = NULL)

Arguments

rows

Numeric. The number of rows the plate should have.

cols

Numeric. The number of columns the plate should have.

wells

Numeric. The number of wells the plate has. If this is specified, rows and cols must be null - they are inferred from common form factors of plates.

Value

a gp object

Details

A gp object has the following components:

nrow/ncol: Number of plate rows/cols. This is static and will not be changed by adding layers.

wells: Number of plate wells. Static.

wells_sec: Number of wells in the given section. When creating a plate, the section refers to the whole plate. This refers to the number of wells without the margin.

wells_sec_par: Number of wells in the section of the parent layer. When creating a plate, there is no parent layer, so the plate acts as its own parent.

well_data: Somewhat transient data used to define plotting coordinates for layers. See below for more information.

nrow_sec/ncols_sec: The number of rows/cols of the current section. When creating a plate, that number is the number of rows/cols of the plate (the plate is the section).

nrow_sec_par/ncol_sec_par: The number of rows/cols of the parent section. When creating a plate, it has no parent, so defaults to being its own parent.

well_data consists of many columns. The variable names can be broken down as follows:

row/col: row is always the y axis, col is always the x axis. By convention, plates start at 1, 1 in the top left corner.

sec: Short for 'section'. A section is a rectangular field of wells. sec alone refers to the number of the section itself. sec combined with row or col (eg row_sec) refers to the coordinates of a given well relative to it's section corners, with the top left corner of a given section always being (1, 1).

rel: Short for 'relative'. This means that the numbering is relative to the starting_corner argument used to create the given section. For instance, a section created with a starting_corner of br (bottom right) will have a row_sec_rel and col_sec_rel of 1 in each of its bottom right corners. A column labelled row_rel in this example would refer to the bottom right of the plate as 1, but not necessarily the bottom right of each section.

par: Short for 'parent'. These columns are all the data from the previous layer.

lane: These specify (usually) multiwell strips only defined in one direction. lane_h and lane_v in tandem from checkerboard-like patterns, where each intersection is a section. This is a bit more complicated when wrap = TRUE, so the simile does not hold for all cases.

col/row_rel_child_sec_par: These are odd columns. They represent the axes of parental sections, but using the child's relative coordinate system. For example, if a parent defined sections starting in the upper left, and the child defined sections starting in the upper right, col_rel_child_sec_par would be a reversed col_sec_par_rel, while row_rel_child_sec_par would be identical to row_sec_par_rel

Examples

# If you specify wells, rows and columns are derived from a standard plate sizes:

gp(wells = 96)
#> Error in loadNamespace(x): there is no package called ‘tidyr’

# As such, you cannot use the wells argument if you want to create more exotic plates:

try(gp(wells = 102))
#> Error in gp(wells = 102) : 
#>   wells is not of a known format (6, 12, 24, 48, 96, 384, 1536, or 3456)

# For that, you'll need to specify wells and cols:

gp(rows = 6, cols = 17)
#> Error in loadNamespace(x): there is no package called ‘tidyr’