Predict from a fitted Gaussian process model

Predict from a fitted Gaussian process model. Predictions are done at the level of the Gaussian Process, i.e. along the factor specified by gp$GP_factor.

# S3 method for class 'gp'
predict(
  gp,
  newdata = NULL,
  hyperpar = gpHyperparList(gp),
  components = NULL,
  comp_missing = c("avg", "none"),
  w = NULL,
  groupMeans = NULL,
  type = c("latent", "response"),
  se.fit = FALSE,
  cov.fit = FALSE,
  CI = 0.95,
  link = NULL,
  parname = NULL,
  maxn = NULL,
  pred.sims = 1e+05,
  Kx.cache = NULL,
  Kxx.cache = NULL,
  ...
)

Arguments

gp

object of class gp, the fitted Gaussian process model

newdata

object of class gpData, unscaled (i.e. not passed through gpDataPrepare()) data! They must be on the same scale as the unscaled training dataset used for fitting the model. If NULL (the default), predictions will be made on the training dataset (which is going to be super fast if CI = NULL)

hyperpar

(optional) lists of hyperparameter values, as returned by gpHyperparList(); default is to use the hyperparameters from the gp object.

components

of the covariance formula. One of the following:

NULL : all components will be selected (the default)

character vector: components to be selected

formula: additive formula specifying components, same syntax as in update.formula, e.g.

~. 	all components

~.-env-spat 	all components except env and spat

~spat+year 	only components spat and year

comp_missing

how to compensate for missing components (if there are any), in mean and variance of the predictions:

"none"

no compensation. In this case, the interpretation of the result's variance isn't much clear (!!!??? why?).

"avg"

the missing components are replaced by "averaging over them" - more precisely, it predicts for the average effects of these components - for these, the variance/CI corresponds to confidence interval (uncertainty around the mean predictions) and not the prediction interval - see http://www.sthda.com/english/articles/40-regression-analysis/166-predict-in-r-model-predictions-and-confidence-intervals/ For models which have grouping factors, be aware of how the average is weighted - see CAUTION CAU001 in the code.

Thus, the most reasonable setting now is "avg"; unless one cares only about the prediction mean (f) of the latent variable, in which case a simple "none" would also do the job.

w

providing this optional argument will calculate average prediction(s) of all data points in this dataset, given these weights. Can be either a vector of the same length as the size of newdata (in which case a single average is calculated), or, if more different averages with different weights are desired, w can be a matrix where each column corresponds to one average - one set of weights (i.e. number of rows of w must be equal to the size of newdata, number of columns equal to the number of different averages to be calculated).

groupMeans

optional, factor of the size of the predicted dataset. If provided, mean predictions for each factor level are calculated.

type

character; type of prediction:

"response"

(default) - predictions on the response scale

"latent"

predictions on the latent scale

cov.fit

logical; if TRUE, the covariance matrix of the predictions will also be returned (can be very memory consuming for large datasets!). Default is FALSE.

link

character, optional. In case of type = "response", what should be the link function to take inverse of for calculating the derived quantity (response scale)? Character as passed to make.link().

parname

character, optional. In case of type = "response", what should be the name of the parameter which is the derived quantity on the response scale?

maxn

maximum maximum dataset size to fit at once (!!! the unit ???!!!); if the dataset is larger, it will be split into chunks of size maxn before predictions are calculated. This parameter doesn't affect the result. It is only to speed-up the computation and use less memory by splitting the dataset

the K(newdata,newdata) matrix is then smaller. So it only speeds it up for CI != NULL (in the backsolve function) for CI = NULL its only slowdown.

pred.sims

(currently unused) number of simulations to be used for calculating the response scale predictions and their CIs. Default is 100000. Higher values give more accurate results, but are slower. Ignored if CI = NULL.

Kx.cache

optional, object returned by K_cache() function, to speed up repeated calls to predict.gp() with the same newdata by caching parts of the K(training_data, newdata) matrix.

Kxx.cache

optional, object returned by K_cache() function, to speed up repeated calls to predict.gp() with the same newdata by caching parts of the K(newdata, newdata) matrix.

!!!!!!

upravit po nove definici - mozna udelat default NULL? CI numeric between 0 and 1, or NULL; confidence level for the confidence intervals to be calculated. If NULL, no standard errors and confidence intervals will be calculated, which makes the prediction much faster (but then also no mean on the response scale will be calculated - because for this, the standard error of the latent GP f is needed.)). Default is 0.95.